Created API controllers and updates the services to work with them properly. Closes #17

+++ src/main/java/edu/greenriver/sdev/myspringproject/services/JobService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import edu.greenriver.sdev.myspringproject.db.IJobRepo;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Random;
 
 /**
@@ -31,9 +32,10 @@ public class JobService {
 	 * saves a job object to the database
 	 * CREATE/UPDATE
 	 * @param job a job object to save
+	 * @return a Job object for the API controller
 	 */
-	public void saveJob(Job job) {
-		jobRepo.save(job);
+	public Job saveJob(Job job) {
+		return jobRepo.save(job);
 	}
 
 	/**
@@ -71,7 +73,33 @@ public class JobService {
 	 * @param id id of record in DB to delete
 	 */
 	public void deleteJob(int id) {
+		if (jobRepo.findById(id).isEmpty()) {
+			throw new NoSuchElementException("Missing Job");
+		}
 		jobRepo.deleteById(id);
+	}
+
+	/**
+	 * Tells us if the job exists in the DB
+	 *
+	 * @param id ID to find Job by
+	 * @return true if the job exists in the DB, otherwise false
+	 */
+	public boolean jobExists(int id) {
+		return oneJob(id) == null;
+	}
+
+	/**
+	 * Edits the Job in the DB
+	 *
+	 * @param job Job to edit
+	 * @return the job to save, for the API controller
+	 */
+	public Job editJob(Job job) {
+		if (jobRepo.findById(job.getJobID()).isEmpty()) {
+			throw new NoSuchElementException("Missing product");
+		}
+		return jobRepo.save(job);
 	}
 
 	@Override
@@ -80,4 +108,5 @@ public class JobService {
 				"jobRepo=" + jobRepo +
 				'}';
 	}
+
 }
